Implement JWT authentication with a cookie
It seems to be working fine. The third-party cookies problem
experienced in the previous project should not exist by creating a
cookie on the frontend instead of on the backend. I also hope that I'll
be able to make the app resistant to browser refreshes, which often
caused my Messaging App project to crash - so far, so good here.