Shrink header on scroll
I also improved the CSS of both the header and the footer. In
particular, I changed how the header looks on mobile.
Also, there is a flickering header bug on very slow scroll near the
specified scroll value. Unfortunately, AFAIK, there is no easy fix for
it, so I decided to leave it as it is.